feature - navigate to 404 on no access. Display banner if app not initialized

+++ src/backend/admin-group-use-case.js
@@ -1,11 +1,10 @@
 import { createActiveAdminGroup } from "../core/db/admin-groups-repository";
 import { generateSortableId } from "../core/db/sortable-id";
-import { deleteActiveAdminGroup, fetchAllAdminGroups } from "./db/admin-group-repostiory";
+import { deleteActiveAdminGroup, fetchAdminGroup, fetchAllAdminGroups } from "./db/admin-group-repostiory";
 import api, { route } from "@forge/api";
 
 export const fetchAdminGroups = async () => {
   const dbData = await fetchAllAdminGroups();
-  console.log("Fetched admin groups:", dbData);
   return dbData.results.map((item) => ({
     ...item.value,
     id: item.key,
@@ -19,16 +18,33 @@ export const addAdminGroups = async (context) => {
   for (const group of adminGroups) {
     promises.push(createActiveAdminGroup(generateSortableId(), { name: group.label, canBeDeleted: true }));
   }
-
   await Promise.all(promises);
 };
 
 export const deleteAdminGroup = async (context) => {
   const adminGroupId = context.payload.id;
-  await deleteActiveAdminGroup(adminGroupId);
+  const rowToDelete = await fetchAdminGroup(adminGroupId);
+  console.log("Deleting admin group:", rowToDelete);
+  if (rowToDelete.value?.canBeDeleted) {
+    await deleteActiveAdminGroup(adminGroupId);
+  }
+};
+
+export const hasUserAdminAccess = async (site) => {
+  // carefull -> it return only 200 first groups
+  const response = await api.asApp().requestConfluence(route`/wiki/rest/api/user/memberof?accountId=${site.context.accountId}`, {
+    headers: {
+      Accept: "application/json",
+    },
+  });
+
+  const userGroups = await response.json();
+  const dbAdminGroups = await fetchAdminGroups();
+  return userGroups.results.some((userGroup) => dbAdminGroups.some((adminGroup) => adminGroup.name === userGroup.name));
 };
 
 export const fetchAllAttlasianGroups = async () => {
+  // carefull -> it return only 200 first groups
   const response = await api.asApp().requestConfluence(route`/wiki/rest/api/group`, {
     headers: {
       Accept: "application/json",
@@ -40,4 +56,8 @@ export const fetchAllAttlasianGroups = async () => {
     label: item.name,
     value: item.name,
   }));
+};
+
+export const verifyIfAppInitialized = async () => {
+  return (await fetchAllAdminGroups()).results.length;
 };
